DeploymentFiles_On_Heuko/requirements.txt
tensorflow-cpu
gunicorn
Flask
opencv-python-headless
Pillow